a ton of restructuring

+++ src/main/scala/numerics/package.scala
@@ -11,14 +11,14 @@ object fun {
    * If base^exponent doesn't fit in a Long, the result will overflow (unlike
    * scala.math.pow which will return +/- Infinity). 
    */
-  final def pow(base:Long, exponent:Long):Long = if (exponent < 0L) {
+  final def pow(base:Long, exponent:Int):Long = if (exponent < 0L) {
     _inv_pow(base, exponent)
   } else {
     _pow(1L, base, exponent)
   }
 
   // inverse powers for integers will return -1L, 0L, 1L, or throw an error.
-  private final def _inv_pow(base:Long, exponent:Long) = if(base == 0L) {
+  private final def _inv_pow(base:Long, exponent:Int) = if(base == 0L) {
     throw new Exception("zero can't be raised to negative power")
   } else if (base == 1L) {
     1L
@@ -29,9 +29,9 @@ object fun {
   }
 
   // tail-recursive helper method for pow(Long, Long)
-  @tailrec private final def _pow(t:Long, b:Long, e:Long): Long = {
-    if (e == 0L) return t
-    _pow(if (e % 2 == 1L) t * b else t, b * b, e / 2L)
+  @tailrec private final def _pow(t:Long, b:Long, e:Int): Long = {
+    if (e == 0) return t
+    _pow(if (e % 2 == 1) t * b else t, b * b, e / 2)
   }
 
   final def pow(base:Double, exponent:Double):Double = {
